refactor(fetcher): extract types to make src/fetcher/ standalone
- Move GetContent, GetContentBytesLength, RequestConfig from
  EpubNavigator.ts to src/fetcher/types.ts
- ContentFetcher and HttpFetcher import from ./types instead of navigator
- EpubNavigator re-exports for backwards compatibility
- src/fetcher/ now has zero navigator dependencies
